refactor: move horizontal bar chart to shared components
