+++ src/app/shared/components/horizontal-bar-chart/horizontal-bar-chart.component.html
